// pharmacy-ui/src/app/services/models/product-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductCategoryResponse } from '../models/product-category-response';
export interface ProductResponse {
  category?: ProductCategoryResponse;
  description?: string;
  discount?: number;
  expiryDate?: string;
  image?: string;
  manufacturer?: string;
  name?: string;
  newArrival?: boolean;
  originalPrice?: number;
  price?: number;
  productId?: number;
  stockQuantity?: number;
  trending?: boolean;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
}